Rewrite based on LoweredCodeUtils.CodeEdges
This is a major rewrite of the backedge component.
Moving it to LoweredCodeUtils and giving it
a "real" API led to many improvements
in selectivity of code-evaluation.

This rewrite aims to make Revise do just one thing,
which is to update methods. "Data" is not touched.
In this world, #249 is not a bug.
The benefit is that `includet` should work much more
reliably to distinguish "code" from "method definitions";
by and large, the days of needing to separate code
from method definitions in order to avoid confusing
Revise should be over.

Importantly, this introduces multiple "modes" of
operation. `:sigs` is for signature extraction,
`:eval` is for full evaluation (as you might do
during the initial `includet` on a file), and
`:evalmeth` is used during revision. This is essentially
a generalization of `define::Bool` to allow more
nuanced behavior.

Fixes #479
Closes #249