Evaluate expressions incrementally in `includet` (fixes #475)
In `includet` mode you want to evaluate each statement.
However, when the same expression is repeated more than once,
the Dict key is identical and thus Revise keeps track of only one of
the expressions. This fixes the problem by evaluating every expression
as it is encountered. The work is done via a modified version
of JuliaInterpreter.split_expressions.

This also significantly changes how expression-wrapping is done.
This new style should preserve more line numbers in the early
phases, though of course the `unwrap` calls end up discarding that
information. But this at least gives the option of keeping more
source info, which may improve backtraces, etc.